+++ pkg/privatelinks/scope_test.go
@@ -5,9 +5,10 @@ import (
 
 	. "github.com/onsi/ginkgo/v2"
 	. "github.com/onsi/gomega"
+	corev1 "k8s.io/api/core/v1"
+	"k8s.io/apimachinery/pkg/runtime"
 	capz "sigs.k8s.io/cluster-api-provider-azure/api/v1beta1"
-	capi "sigs.k8s.io/cluster-api/api/v1beta1"
-	"sigs.k8s.io/cluster-api/util/conditions"
+	capi "sigs.k8s.io/cluster-api/api/core/v1beta1"
 	"sigs.k8s.io/controller-runtime/pkg/client"
 	"sigs.k8s.io/controller-runtime/pkg/client/fake"
 
@@ -21,10 +22,13 @@ var linkName = "test-private-link"
 var _ = Describe("Scope", func() {
 	var subscriptionID string
 	var resourceGroup string
+	var scheme *runtime.Scheme
 
 	BeforeEach(func() {
 		subscriptionID = "1234"
 		resourceGroup = "test-rg"
+		scheme = runtime.NewScheme()
+		Expect(capz.AddToScheme(scheme)).To(Succeed())
 	})
 
 	Describe("creating scope", func() {
@@ -32,11 +36,12 @@ var _ = Describe("Scope", func() {
 		var client client.Client
 
 		BeforeEach(func() {
-			azureCluster = testhelpers.NewAzureClusterBuilder(subscriptionID, resourceGroup).Build()
-			capzSchema, err := capz.SchemeBuilder.Build()
-			Expect(err).NotTo(HaveOccurred())
+			azureCluster = testhelpers.NewAzureClusterBuilder("org-giantswarm", resourceGroup).
+				WithSubscriptionID(subscriptionID).
+				WithResourceGroup(resourceGroup).
+				Build()
 			client = fake.NewClientBuilder().
-				WithScheme(capzSchema).
+				WithScheme(scheme).
 				WithObjects(azureCluster).Build()
 		})
 
@@ -69,16 +74,18 @@ var _ = Describe("Scope", func() {
 		var scope *privatelinks.Scope
 
 		JustBeforeEach(func() {
-			azureClusterBuilder := testhelpers.NewAzureClusterBuilder(subscriptionID, resourceGroup)
+			azureClusterBuilder := testhelpers.NewAzureClusterBuilder("org-giantswarm", resourceGroup).
+				WithSubscriptionID(subscriptionID).
+				WithResourceGroup(resourceGroup)
+
 			for _, privateLinkName := range privateLinkNames {
 				azureClusterBuilder.WithPrivateLink(testhelpers.NewPrivateLinkBuilder(*privateLinkName).Build())
 			}
 			azureCluster := azureClusterBuilder.Build()
-			capzSchema, err := capz.SchemeBuilder.Build()
-			Expect(err).NotTo(HaveOccurred())
 			client := fake.NewClientBuilder().
-				WithScheme(capzSchema).
+				WithScheme(scheme).
 				WithObjects(azureCluster).Build()
+			var err error
 			scope, err = privatelinks.NewScope(azureCluster, client)
 			Expect(err).NotTo(HaveOccurred())
 		})
@@ -142,7 +149,10 @@ var _ = Describe("Scope", func() {
 		var scope *privatelinks.Scope
 
 		JustBeforeEach(func() {
-			azureClusterBuilder := testhelpers.NewAzureClusterBuilder(subscriptionID, resourceGroup)
+			azureClusterBuilder := testhelpers.NewAzureClusterBuilder("org-giantswarm", resourceGroup).
+				WithSubscriptionID(subscriptionID).
+				WithResourceGroup(resourceGroup)
+
 			for privateLinkName, allowedSubscriptions := range privateLinksWithAllowedSubscriptions {
 				privateLinkBuilder := testhelpers.NewPrivateLinkBuilder(privateLinkName)
 				for _, allowedSubscription := range allowedSubscriptions {
@@ -152,11 +162,10 @@ var _ = Describe("Scope", func() {
 				azureClusterBuilder.WithPrivateLink(privateLink)
 			}
 			azureCluster := azureClusterBuilder.Build()
-			capzSchema, err := capz.SchemeBuilder.Build()
-			Expect(err).NotTo(HaveOccurred())
 			client := fake.NewClientBuilder().
-				WithScheme(capzSchema).
+				WithScheme(scheme).
 				WithObjects(azureCluster).Build()
+			var err error
 			scope, err = privatelinks.NewScope(azureCluster, client)
 			Expect(err).NotTo(HaveOccurred())
 		})
@@ -213,23 +222,27 @@ var _ = Describe("Scope", func() {
 		var scope *privatelinks.Scope
 
 		JustBeforeEach(func() {
-			azureCluster := testhelpers.NewAzureClusterBuilder(subscriptionID, resourceGroup).
+			azureCluster := testhelpers.NewAzureClusterBuilder("org-giantswarm", resourceGroup).
+				WithSubscriptionID(subscriptionID).
+				WithResourceGroup(resourceGroup).
 				WithPrivateLink(testhelpers.NewPrivateLinkBuilder("test-private-link").Build()).
 				WithCondition(privateLinksCondition).
 				Build()
 
-			capzSchema, err := capz.SchemeBuilder.Build()
-			Expect(err).NotTo(HaveOccurred())
 			client := fake.NewClientBuilder().
-				WithScheme(capzSchema).
+				WithScheme(scheme).
 				WithObjects(azureCluster).Build()
+			var err error
 			scope, err = privatelinks.NewScope(azureCluster, client)
 			Expect(err).NotTo(HaveOccurred())
 		})
 
 		When("PrivateLinksReady condition has status True", func() {
 			BeforeEach(func() {
-				privateLinksCondition = conditions.TrueCondition(capz.PrivateLinksReadyCondition)
+				privateLinksCondition = &capi.Condition{
+					Type:   capz.PrivateLinksReadyCondition,
+					Status: corev1.ConditionTrue,
+				}
 			})
 			It("scope.PrivateLinksReady returns true", func() {
 				privateLinksReady := scope.PrivateLinksReady()
@@ -239,7 +252,11 @@ var _ = Describe("Scope", func() {
 
 		When("PrivateLinksReady condition has status False", func() {
 			BeforeEach(func() {
-				privateLinksCondition = conditions.FalseCondition(capz.PrivateLinksReadyCondition, "Something", capi.ConditionSeverityError, "some error")
+				privateLinksCondition = &capi.Condition{
+					Type:   capz.PrivateLinksReadyCondition,
+					Status: corev1.ConditionFalse,
+					// Severity: capi.ConditionSeverityError,
+				}
 			})
 			It("scope.PrivateLinksReady returns false", func() {
 				privateLinksReady := scope.PrivateLinksReady()
@@ -249,7 +266,10 @@ var _ = Describe("Scope", func() {
 
 		When("PrivateLinksReady condition has status Unknown", func() {
 			BeforeEach(func() {
-				privateLinksCondition = conditions.UnknownCondition(capz.PrivateLinksReadyCondition, "Something", "some error")
+				privateLinksCondition = &capi.Condition{
+					Type:   capz.PrivateLinksReadyCondition,
+					Status: corev1.ConditionUnknown,
+				}
 			})
 			It("scope.PrivateLinksReady returns false", func() {
 				privateLinksReady := scope.PrivateLinksReady()
@@ -273,12 +293,16 @@ var _ = Describe("Scope", func() {
 		var scope *privatelinks.Scope
 
 		BeforeEach(func() {
-			azureCluster = testhelpers.NewAzureClusterBuilder(subscriptionID, resourceGroup).Build()
-			capzSchema, err := capz.SchemeBuilder.Build()
-			Expect(err).NotTo(HaveOccurred())
+			azureCluster = testhelpers.NewAzureClusterBuilder("org-giantswarm", resourceGroup).
+				WithSubscriptionID(subscriptionID).
+				WithResourceGroup(resourceGroup).
+				Build()
+
 			client := fake.NewClientBuilder().
-				WithScheme(capzSchema).
+				WithScheme(scheme).
 				WithObjects(azureCluster).Build()
+
+			var err error
 			scope, err = privatelinks.NewScope(azureCluster, client)
 			Expect(err).NotTo(HaveOccurred())
 		})