+++ pkg/testhelpers/cluster.go
@@ -1,40 +1,56 @@
 package testhelpers
 
 import (
-	v1 "k8s.io/api/core/v1"
-	"k8s.io/apimachinery/pkg/runtime"
+	"time"
+
+	meta "k8s.io/apimachinery/pkg/apis/meta/v1"
 	capz "sigs.k8s.io/cluster-api-provider-azure/api/v1beta1"
-	clusterv1 "sigs.k8s.io/cluster-api/api/v1beta1"
-	kcpv1 "sigs.k8s.io/cluster-api/controlplane/kubeadm/api/v1beta1"
+	kcp "sigs.k8s.io/cluster-api/api/controlplane/kubeadm/v1beta2"
+	"sigs.k8s.io/cluster-api/api/core/v1beta2"
+	capi "sigs.k8s.io/cluster-api/api/core/v1beta2"
 	ctrl "sigs.k8s.io/controller-runtime"
+	"sigs.k8s.io/controller-runtime/pkg/client/apiutil"
 )
 
-func NewClusterBuilder(scheme *runtime.Scheme) *ClusterBuilder {
-	return &ClusterBuilder{
-		o:      new(clusterv1.Cluster),
-		scheme: scheme,
-	}
+func NewClusterBuilder(namespace, name string) *ClusterBuilder {
+	ensureSchemeSet()
+
+	o := new(capi.Cluster)
+	o.SetNamespace(namespace)
+	o.SetName(name)
+	return &ClusterBuilder{o}
 }
 
 type ClusterBuilder struct {
-	o      *clusterv1.Cluster
-	scheme *runtime.Scheme
+	o *capi.Cluster
 }
 
-func (b *ClusterBuilder) WithPause() *ClusterBuilder {
-	b.o.Spec.Paused = true
+func (b *ClusterBuilder) WithFinalizers(finalizers ...string) *ClusterBuilder {
+	if b.o.Finalizers == nil {
+		b.o.Finalizers = make([]string, 0)
+	}
+	b.o.Finalizers = append(b.o.Finalizers, finalizers...)
 	return b
 }
 
-func (b *ClusterBuilder) WithControlPlane(kcp *kcpv1.KubeadmControlPlane) *ClusterBuilder {
-	b.o.ObjectMeta.Namespace = kcp.Namespace
-	b.o.ObjectMeta.Name = kcp.Name
-	b.o.Spec.ControlPlaneRef = &v1.ObjectReference{
-		Kind:      kcp.Kind,
-		Namespace: kcp.Namespace,
-		Name:      kcp.Name,
+func (b *ClusterBuilder) WithDeletionTimestamp(time time.Time) *ClusterBuilder {
+	ts := meta.NewTime(time)
+	b.o.DeletionTimestamp = &ts
+	return b
+}
+
+func (b *ClusterBuilder) WithPause() *ClusterBuilder {
+	b.o.Spec.Paused = new(true)
+	return b
+}
+
+func (b *ClusterBuilder) WithControlPlane(kcp *kcp.KubeadmControlPlane) *ClusterBuilder {
+	b.o.Spec.ControlPlaneRef = capi.ContractVersionedObjectReference{
+		APIGroup: capi.GroupVersionControlPlane.Group,
+		Kind:     kcp.Kind,
+		Name:     kcp.Name,
 	}
-	err := ctrl.SetControllerReference(b.o, kcp, b.scheme)
+	err := ctrl.SetControllerReference(b.o, kcp, scheme)
 	if err != nil {
 		panic(err)
 	}
@@ -42,16 +58,39 @@ func (b *ClusterBuilder) WithControlPlane(kcp *kcpv1.KubeadmControlPlane) *Clust
 }
 
 func (b *ClusterBuilder) WithAzureCluster(ac *capz.AzureCluster) *ClusterBuilder {
-	b.o.Namespace = ac.Namespace
-	b.o.Name = ac.Name
-	b.o.Spec.InfrastructureRef = &v1.ObjectReference{
-		Kind:      ac.Kind,
-		Namespace: ac.Namespace,
-		Name:      ac.Name,
+	b.o.Spec.InfrastructureRef = capi.ContractVersionedObjectReference{
+		APIGroup: capi.GroupVersionInfrastructure.Group,
+		Kind:     ac.Kind,
+		Name:     ac.Name,
 	}
 	return b
 }
 
-func (b *ClusterBuilder) Build() *clusterv1.Cluster {
+// WithDummyReferences populates the Cluster with dummy controlplane and infrastructure references.
+// This is useful for when you need only need a Cluster object that does not need to point
+// to real controlplane or infrastructure objects.
+// If you need a dummy controlplane but not a dummy infrastructure reference, or vice versa,
+// you can call this method and then afterwards call a method to set a reference to the real object
+// that you need.
+func (b *ClusterBuilder) WithDummyReferences() *ClusterBuilder {
+	b.o.Spec.InfrastructureRef = v1beta2.ContractVersionedObjectReference{
+		APIGroup: "infrastructure.cluster.x-k8s.io",
+		Kind:     "DummyInfraCluster",
+		Name:     "dummy-infracluster",
+	}
+	b.o.Spec.ControlPlaneRef = v1beta2.ContractVersionedObjectReference{
+		APIGroup: "controlplane.cluster.x-k8s.io",
+		Kind:     "DummyControlPlane",
+		Name:     "dummy-controlplane",
+	}
+	return b
+}
+
+func (b *ClusterBuilder) Build() *capi.Cluster {
+	gvk, err := apiutil.GVKForObject(b.o, scheme)
+	if err != nil {
+		panic(err)
+	}
+	b.o.SetGroupVersionKind(gvk)
 	return b.o
 }