+++ CHANGELOG.md
@@ -7,6 +7,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+- Update CAPI dependency from v1.9.4 to v1.12.4.
+- Update CAPZ dependency from 1.12.4 to v1.23.0.
+- Update controller-runtime from v0.19.4 to v0.22.5.
+
 ## [0.5.1] - 2026-05-19
 
 ### Changed